image/test/mochitest/test_bug657191.html

<!DOCTYPE HTML>
<html>



<head>
  <title>Test for Bug 657191</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=657191">Mozilla Bug 657191</a>
<p id="display"></p>
<div id="content" style="display: none">


  <img src="bug657191.sjs">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 657191 **/
SimpleTest.waitForExplicitFinish();

window.onload = function() {
  ok(true, "test passed");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
